move cloudcore apis to pkg dir, move edgecore apis to pkg dir
