+++ cloud/cmd/cloudcore/app/server.go
@@ -13,11 +13,11 @@ import (
 
 	"github.com/kubeedge/beehive/pkg/core"
 	"github.com/kubeedge/kubeedge/cloud/cmd/cloudcore/app/options"
-	"github.com/kubeedge/kubeedge/cloud/pkg/apis/cloudcore/config"
-	"github.com/kubeedge/kubeedge/cloud/pkg/apis/cloudcore/config/validation"
 	"github.com/kubeedge/kubeedge/cloud/pkg/cloudhub"
 	"github.com/kubeedge/kubeedge/cloud/pkg/devicecontroller"
 	"github.com/kubeedge/kubeedge/cloud/pkg/edgecontroller"
+	"github.com/kubeedge/kubeedge/pkg/cloudcore/apis/config"
+	"github.com/kubeedge/kubeedge/pkg/cloudcore/apis/config/validation"
 	"github.com/kubeedge/kubeedge/pkg/util/flag"
 	"github.com/kubeedge/kubeedge/pkg/version"
 	"github.com/kubeedge/kubeedge/pkg/version/verflag"
