+++ pkg/cloudcore/apis/config/validation/validation_test.go
